Rename: pages 하위 폴더명 수정
