+++ src/pages/joinPage/JoinStyle.jsx
