+++ src/pages/mainPage/MainStyle.jsx
